Fixed getCredentials and performlogin (#22)
When working on another feature i discovered that the performlogin and getcredentials function in auth0repo didn't work as they were supposed to. I fixed this by having both function return flows and working with the credentialsManager to save and get the credentials. Both functions also work with the ApiResource to handle state and emit correctly.