WIP backend implementation
This branch contains added backend logic structure to add a new tenturncard including a testcase for the dao object. This testcase passes but to run it succesfully other broken testcases need to be placed in comments otherwise you get a compile error (this needs to be fixed ASAP)

In the commit following this one I will also add a testcase for the tenturncardrepo and afterwards this can be merged with main so i can create a  seperate branch for implementing the ui side of things.