Backend
This commit contains the "backend" logic behind updating a tenturncard that, in a future commit, the viewmodel will use.

The necesarry testcases have also been updated to test this new logic.

I also noticed that the addTenturncard function didn't have the suspend modifiier which I thought was strange so I fixed this.